refactor(app): inline object request context types
